+++ admin/src/scenes/dashboardV2/components/inscription/TabSchool.js
@@ -1,10 +1,11 @@
 import React, { useEffect } from "react";
-import { HiOutlineChevronLeft, HiOutlineChevronRight } from "react-icons/hi";
+import { HiOutlineChevronLeft, HiOutlineChevronRight, HiOutlineExternalLink } from "react-icons/hi";
 import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { ROLES } from "snu-lib";
 import api from "../../../../services/api";
 import { replaceSpaces } from "../../../../utils";
+import { currentFilterAsUrl } from "../FilterDashBoard";
 
 const PAGE_SIZE = 6;
 
@@ -38,6 +39,7 @@ export default function TabSchool({ filters }) {
     if (filters.region?.length) body.query.bool.filter.push({ terms: { "schoolRegion.keyword": filters.region } });
     if (filters.department?.length) body.query.bool.filter.push({ terms: { "schoolDepartment.keyword": filters.department } });
     if (filters.cohort?.length) body.query.bool.filter.push({ terms: { "cohort.keyword": filters.cohort } });
+    if (filters.academy?.length) body.query.bool.filter.push({ terms: { "academy.keyword": filters.academy } });
 
     let route = "young";
     if (user.role === ROLES.REFERENT_DEPARTMENT) route = "young-having-school-in-department/inscriptions";
@@ -87,9 +89,14 @@ export default function TabSchool({ filters }) {
   }, [filters]);
 
   return (
-    <div className="flex flex-col gap-5 w-[60%] bg-white rounded-lg px-8 py-8">
+    <div className="flex flex-col gap-5 w-[60%] bg-white rounded-lg px-8 py-8 shadow-[0_8px_16px_-3px_rgba(0,0,0,0.05)]">
       <div className="flex flex-row justify-between w-full">
-        <div className="text-base font-bold text-gray-900">Liste des établissements</div>
+        <div className="flex items-center gap-3">
+          <div className="text-base font-bold text-gray-900">Liste des établissements</div>
+          <Link to={`/etablissement/liste-jeunes?${currentFilterAsUrl(filters)}`} target={"_blank"}>
+            <HiOutlineExternalLink className="h-5 w-5 text-gray-400 cursor-pointer" />
+          </Link>
+        </div>
         <div className="text-xs text-gray-600">
           Export depuis le menu{" "}
           <Link
@@ -106,33 +113,33 @@ export default function TabSchool({ filters }) {
       <table className={`table-fixed w-full ${isLoading || noResult ? "h-full" : ""}`}>
         <thead>
           <tr className="flex items-center border-y-[1px] border-gray-100 py-4">
-            <th className="w-[70%] uppercase text-xs text-gray-500 font-medium leading-4">établissements</th>
-            <th className="w-[30%] uppercase text-xs text-gray-500 font-medium leading-4">volontaires</th>
+            <th className="w-[80%] uppercase text-xs text-gray-400 font-medium leading-4">établissements</th>
+            <th className="w-[20%] uppercase text-xs text-gray-400 font-medium leading-4">volontaires</th>
           </tr>
         </thead>
         <tbody className="">
           {isLoading ? (
             Array.from(Array(PAGE_SIZE).keys()).map((i) => (
               <tr key={`LoadingSchool${i}`} className="flex items-center border-b-[1px] border-gray-100 h-1/6">
-                <td className="w-[70%]">
+                <td className="w-[80%]">
                   <Loading width="w-[50%]" />
                 </td>
-                <td className="w-[30%]">
+                <td className="w-[20%]">
                   <Loading width="w-[50%]" />
                 </td>
               </tr>
             ))
           ) : !noResult ? (
             youngBySchool?.slice(page * PAGE_SIZE, page * PAGE_SIZE + PAGE_SIZE)?.map((school) => (
-              <tr key={school?.key} className="flex items-center border-b-[1px] border-gray-100 py-3 h-1/6">
-                <td className="flex flex-col w-[70%] gap-1">
+              <tr key={school?.key} className="flex items-center border-b-[1px] border-gray-100 py-3 h-1/6 hover:bg-gray-50 cursor-default">
+                <td className="flex flex-col w-[80%] gap-1">
                   <Link
                     to={`/inscription?SCHOOL=%5B"${replaceSpaces(school.schoolName)}"%5D`}
                     target="_blank"
                     onClick={(e) => {
                       e.stopPropagation();
                     }}
-                    className="text-blue-600 cursor-pointer">
+                    className="cursor-pointer">
                     <p className="text-sm text-gray-900 font-bold leading-6 truncate w-[90%]">{school.schoolName}</p>
                     <p className="text-xs text-gray-500 leading-4">
                       {school?.schoolCity} • {school.schoolZip}
@@ -140,7 +147,7 @@ export default function TabSchool({ filters }) {
                   </Link>
                 </td>
 
-                <td className="w-[30%] flex flex-col gap-1 ">
+                <td className="w-[20%] flex flex-col gap-1 ">
                   <span className="text-sm text-gray-500 leading-3">
                     <span className="text-gray-900 font-bold">{school.total}</span> (dont {school.inDepartment} au sein
                   </span>
@@ -156,7 +163,7 @@ export default function TabSchool({ filters }) {
         </tbody>
       </table>
       <div className="flex items-center justify-between pt-4 w-full">
-        <p className="text-sm leading-5 font-normal text-gray-700">
+        <p className="text-xs leading-5 font-normal text-gray-700">
           {noResult ? 0 : page * PAGE_SIZE + 1}-<strong> {page * PAGE_SIZE + PAGE_SIZE >= total ? total : page * PAGE_SIZE + PAGE_SIZE}</strong> sur <strong>{total}</strong>{" "}
           résultats
         </p>