+++ api/src/controllers/young/index.js
@@ -64,21 +64,6 @@ const { anonymizeApplicationsFromYoungId } = require("../../services/application
 const { anonymizeContractsFromYoungId } = require("../../services/contract");
 const { getFillingRate, FILLING_RATE_LIMIT } = require("../../services/inscription-goal");
 
-const redis = require("redis");
-
-const redisClient = redis.createClient({
-  url: config.REDIS_URL,
-});
-redisClient.connect().catch(console.error);
-
-redisClient.on("connect", function () {
-  console.log("Connected to Redis server");
-});
-
-redisClient.on("error", function (error) {
-  console.error("Error connecting to Redis server", error);
-});
-
 router.post("/signup", (req, res) => YoungAuth.signUp(req, res));
 router.post("/signup2023", (req, res) => YoungAuth.signUp2023(req, res));
 router.post("/signin", (req, res) => YoungAuth.signin(req, res));
@@ -763,19 +748,10 @@ router.post("/france-connect/authorization-url", async (req, res) => {
       redirect_uri: `${config.APP_URL}/${value.callback}`,
       response_type: "code",
       client_id: process.env.FRANCE_CONNECT_CLIENT_ID,
-      state: crypto.randomBytes(20).toString("hex"),
+      state: "home",
       nonce: crypto.randomBytes(20).toString("hex"),
       acr_values: "eidas1",
     };
-
-    try {
-      await redisClient.setEx(`franceConnectNonce:${query.nonce}`, 1800, query.nonce);
-      await redisClient.setEx(`franceConnectState:${query.state}`, 1800, query.state);
-    } catch (e) {
-      capture(e);
-      return res.status(500).send({ ok: false, code: ERRORS.SERVER_ERROR });
-    }
-
     const url = `${process.env.FRANCE_CONNECT_URL}/authorize?${queryString.stringify(query)}`;
     return res.status(200).send({ ok: true, data: { url } });
   } catch (error) {
@@ -787,9 +763,7 @@ router.post("/france-connect/authorization-url", async (req, res) => {
 // Get user information for authorized user on France Connect.
 router.post("/france-connect/user-info", async (req, res) => {
   try {
-    const { error, value } = Joi.object({ code: Joi.string().required(), callback: Joi.string().required(), state: Joi.string().required() })
-      .unknown()
-      .validate(req.body, { stripUnknown: true });
+    const { error, value } = Joi.object({ code: Joi.string().required(), callback: Joi.string().required() }).unknown().validate(req.body, { stripUnknown: true });
     if (error) {
       capture(error);
       return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
@@ -802,31 +776,14 @@ router.post("/france-connect/user-info", async (req, res) => {
       client_secret: process.env.FRANCE_CONNECT_CLIENT_SECRET,
       code: value.code,
     };
-
     const tokenResponse = await fetch(`${process.env.FRANCE_CONNECT_URL}/token`, {
       method: "POST",
       headers: { "content-type": "application/x-www-form-urlencoded" },
       body: queryString.stringify(body),
     });
-
     const token = await tokenResponse.json();
-    const franceConnectToken = token["id_token"];
 
-    const decodedToken = jwt.decode(franceConnectToken);
-
-    let storedState;
-    let storedNonce;
-
-    try {
-      storedState = await redisClient.get(`franceConnectState:${value.state}`);
-      storedNonce = await redisClient.get(`franceConnectNonce:${decodedToken.nonce}`);
-    } catch (e) {
-      capture(e);
-      return res.status(500).send({ ok: false, code: ERRORS.SERVER_ERROR });
-    }
-
-    if (!token["access_token"] || !token["id_token"] || !storedNonce || !storedState) {
-      capture(`France Connect User Information failed: ${JSON.stringify({ storedNonce, storedState, token })}`);
+    if (!token["access_token"] || !token["id_token"]) {
       return res.sendStatus(403, token);
     }
 
@@ -835,9 +792,7 @@ router.post("/france-connect/user-info", async (req, res) => {
       method: "GET",
       headers: { Authorization: `Bearer ${token["access_token"]}` },
     });
-
     const userInfo = await userInfoResponse.json();
-
     res.status(200).send({ ok: true, data: userInfo, tokenId: token["id_token"] });
   } catch (e) {
     capture(e);