+++ api/src/utils/cohort.js
@@ -23,8 +23,8 @@ async function getFilteredSessions(young) {
     (session) =>
       session.eligibility.zones.includes(region2zone[region]) &&
       session.eligibility.schoolLevels.includes(young.grade) &&
-      session.eligibility.bornAfter < young.birthdateAt &&
-      session.eligibility.bornBefore > young.birthdateAt &&
+      session.eligibility.bornAfter <= young.birthdateAt &&
+      session.eligibility.bornBefore >= young.birthdateAt &&
       (session.eligibility.inscriptionEndDate > Date.now() ||
         ([YOUNG_STATUS.WAITING_CORRECTION, YOUNG_STATUS.WAITING_VALIDATION].includes(young.status) && session.eligibility.instructionEndDate > Date.now())),
   );